Stop passport throwing error on undefined user.
Edge case but, should an attempt to log the user in fail, a user ID
will be created in the session, but no record will be returned from the
database (the ID got into the session, but not the DB). This will throw
an error which is displayed to the user, setting the user to null will
allow passport to handle the error gracefully.